requirements.txt
Flask==2.0.1
   twilio==7.8.0
   requests==2.26.0
   openai==0.27.0
   python-dotenv==0.19.0
   gunicorn==20.1.0